refactor(data): centralize database builder
